refactor: move methods to base
