+++ packages/valory/skills/decision_maker_abci/behaviours/base.py
@@ -19,21 +19,34 @@
 
 """This module contains the base behaviour for the 'decision_maker_abci' skill."""
 
+import dataclasses
 from abc import ABC
 from datetime import datetime, timedelta
-from typing import Any, Callable, Generator, Optional, cast
+from typing import Any, Callable, Generator, List, Optional, cast
 
 from aea.configurations.data_types import PublicId
 
+from packages.valory.contracts.gnosis_safe.contract import (
+    GnosisSafeContract,
+    SafeOperation,
+)
 from packages.valory.contracts.mech.contract import Mech
+from packages.valory.contracts.multisend.contract import MultiSendContract
 from packages.valory.protocols.contract_api import ContractApiMessage
 from packages.valory.skills.abstract_round_abci.base import BaseTxPayload
 from packages.valory.skills.abstract_round_abci.behaviour_utils import (
     BaseBehaviour,
     TimeoutException,
 )
-from packages.valory.skills.decision_maker_abci.models import DecisionMakerParams
+from packages.valory.skills.decision_maker_abci.models import (
+    DecisionMakerParams,
+    MultisendBatch,
+)
 from packages.valory.skills.decision_maker_abci.states.base import SynchronizedData
+from packages.valory.skills.transaction_settlement_abci.payload_tools import (
+    hash_payload_to_hex,
+)
+from packages.valory.skills.transaction_settlement_abci.rounds import TX_HASH_LENGTH
 
 
 WaitableConditionType = Generator[None, None, bool]
@@ -48,6 +61,13 @@ SAFE_GAS = 0
 class DecisionMakerBaseBehaviour(BaseBehaviour, ABC):
     """Represents the base class for the decision-making FSM behaviour."""
 
+    def __init__(self, **kwargs: Any) -> None:
+        """Initialize the bet placement behaviour."""
+        super().__init__(**kwargs)
+        self.multisend_batches: List[MultisendBatch] = []
+        self.multisend_data = b""
+        self.safe_tx_hash = ""
+
     @property
     def params(self) -> DecisionMakerParams:
         """Return the params."""
@@ -57,6 +77,38 @@ class DecisionMakerBaseBehaviour(BaseBehaviour, ABC):
     def synchronized_data(self) -> SynchronizedData:
         """Return the synchronized data."""
         return SynchronizedData(super().synchronized_data.db)
+
+    @property
+    def multi_send_txs(self) -> List[dict]:
+        """Get the multisend transactions as a list of dictionaries."""
+        return [dataclasses.asdict(batch) for batch in self.multisend_batches]
+
+    @property
+    def txs_value(self) -> int:
+        """Get the total value of the transactions."""
+        return sum(batch.value for batch in self.multisend_batches)
+
+    @property
+    def tx_hex(self) -> Optional[str]:
+        """Serialize the safe tx to a hex string."""
+        if self.safe_tx_hash == "":
+            self.context.logger.error(
+                "Cannot prepare a transaction without a transaction hash."
+            )
+            return None
+        return hash_payload_to_hex(
+            self.safe_tx_hash,
+            self.txs_value,
+            SAFE_GAS,
+            self.params.multisend_address,
+            self.multisend_data,
+            SafeOperation.DELEGATE_CALL.value,
+        )
+
+    @staticmethod
+    def wei_to_native(wei: int) -> float:
+        """Convert WEI to native token."""
+        return wei / 10**18
 
     def contract_interaction_error(
         self, contract_id: str, contract_callable: str, response_msg: ContractApiMessage
@@ -114,6 +166,72 @@ class DecisionMakerBaseBehaviour(BaseBehaviour, ABC):
             **kwargs,
         )
         return status
+
+    def _build_multisend_data(
+        self,
+    ) -> WaitableConditionType:
+        """Get the multisend tx."""
+        response_msg = yield from self.get_contract_api_response(
+            performative=ContractApiMessage.Performative.GET_RAW_TRANSACTION,  # type: ignore
+            contract_address=self.params.multisend_address,
+            contract_id=str(MultiSendContract.contract_id),
+            contract_callable="get_tx_data",
+            multi_send_txs=self.multi_send_txs,
+        )
+        expected_performative = ContractApiMessage.Performative.RAW_TRANSACTION
+        if response_msg.performative != expected_performative:
+            self.context.logger.error(
+                f"Couldn't compile the multisend tx. "
+                f"Expected response performative {expected_performative.value}, "  # type: ignore
+                f"received {response_msg.performative.value}: {response_msg}"
+            )
+            return False
+
+        multisend_data_str = response_msg.raw_transaction.body.get("data", None)
+        if multisend_data_str is None:
+            self.context.logger.error(
+                f"Something went wrong while trying to prepare the multisend data: {response_msg}"
+            )
+            return False
+
+        # strip "0x" from the response
+        multisend_data_str = str(response_msg.raw_transaction.body["data"])[2:]
+        self.multisend_data = bytes.fromhex(multisend_data_str)
+        return True
+
+    def _build_multisend_safe_tx_hash(self) -> WaitableConditionType:
+        """Prepares and returns the safe tx hash for a multisend tx."""
+        response_msg = yield from self.get_contract_api_response(
+            performative=ContractApiMessage.Performative.GET_STATE,  # type: ignore
+            contract_address=self.synchronized_data.safe_contract_address,
+            contract_id=str(GnosisSafeContract.contract_id),
+            contract_callable="get_raw_safe_transaction_hash",
+            to_address=self.params.multisend_address,
+            value=self.txs_value,
+            data=self.multisend_data,
+            safe_tx_gas=SAFE_GAS,
+            operation=SafeOperation.DELEGATE_CALL.value,
+        )
+
+        if response_msg.performative != ContractApiMessage.Performative.STATE:
+            self.context.logger.error(
+                "Couldn't get safe tx hash. Expected response performative "
+                f"{ContractApiMessage.Performative.STATE.value}, "  # type: ignore
+                f"received {response_msg.performative.value}: {response_msg}."
+            )
+            return False
+
+        tx_hash = response_msg.state.body.get("tx_hash", None)
+        if tx_hash is None or len(tx_hash) != TX_HASH_LENGTH:
+            self.context.logger.error(
+                "Something went wrong while trying to get the buy transaction's hash. "
+                f"Invalid hash {tx_hash!r} was returned."
+            )
+            return False
+
+        # strip "0x" from the response hash
+        self.safe_tx_hash = tx_hash[2:]
+        return True
 
     def wait_for_condition_with_sleep(
         self,
